temp_backup/requirements_temp.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.2
python-dotenv==1.0.0
agno>=1.2.6
openai==1.10.0
anthropic==0.7.7
duckduckgo-search==3.9.3
newspaper3k==0.2.8
python-multipart==0.0.6
requests==2.31.0
python-jose==3.3.0
jinja2==3.1.2
markdown==3.4.4
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
alembic==1.12.0
aiofiles==23.2.1
pytest==7.4.2
pytest-asyncio==0.21.1
boto3==1.28.35
tenacity==8.2.3
starlette==0.46.1
PyYAML==6.0.2
httpx
